refactor: rename RuleToExecute to ExecutableRule
