+++ src/main/scala/com/amazon/deequ/dqdl/EvaluateDataQuality.scala
@@ -16,7 +16,8 @@
 
 package com.amazon.deequ.dqdl
 
-import com.amazon.deequ.dqdl.execution.DeequCheckExecutor
+import com.amazon.deequ.dqdl.execution.DQDLExecutor
+import com.amazon.deequ.dqdl.model.ExecutableRule
 import com.amazon.deequ.dqdl.translation.{DQDLRuleTranslator, DeequOutcomeTranslator}
 import com.amazon.deequ.dqdl.util.DefaultDQDLParser
 import org.apache.spark.sql.DataFrame
@@ -26,8 +27,8 @@ import software.amazon.glue.dqdl.model.DQRuleset
  * Entry point for evaluating data quality.
  *
  * Given a Spark DataFrame and a DQDL ruleset (as a String), this object:
- *  - Parses and translates the rules to Deequ Checks.
- *  - Executes the checks on the DataFrame.
+ *  - Parses and translates each rule from the ruleset to an ExecutableRule
+ *  - Executes the rules on the DataFrame.
  *  - Translates the outcome back to a Spark DataFrame.
  */
 object EvaluateDataQuality {
@@ -38,23 +39,23 @@ object EvaluateDataQuality {
    * <p>This method applies the specified data quality ruleset to the input DataFrame and returns a new
    * DataFrame summarizing the overall quality results, including any issues detected during the analysis.</p>
    *
-   * @param df      the Spark DataFrame to analyze.
-   * @param ruleset the data quality ruleset in DQDL format to apply to the DataFrame.
+   * @param df                the Spark DataFrame to analyze.
+   * @param rulesetDefinition the data quality ruleset (defined in DQDL string format) to apply to the DataFrame.
    * @return a Spark DataFrame containing the aggregated data quality results.
    */
-  def process(df: DataFrame, ruleset: String): DataFrame = {
+  def process(df: DataFrame, rulesetDefinition: String): DataFrame = {
     // 1. Parse the ruleset
-    val dqRuleset: DQRuleset = DefaultDQDLParser.parse(ruleset)
+    val ruleset: DQRuleset = DefaultDQDLParser.parse(rulesetDefinition)
 
-    // 2. Translate the dqRuleset into a corresponding deequRules.
-    val deequRules = DQDLRuleTranslator.toDeequRules(dqRuleset)
+    // 2. Translate the dqRuleset into a corresponding ExecutableRules.
+    val executableRules: Seq[ExecutableRule] = DQDLRuleTranslator.toExecutableRules(ruleset)
 
-    // 3. Execute the checks against the DataFrame.
-    val verificationResult = DeequCheckExecutor.executeChecks(df, Seq.empty)
+    // 3. Execute the rules against the DataFrame.
+    val result = DQDLExecutor.executeRules(df, Seq.empty)
 
-    // 4. Translate the Deequ results into a Spark DataFrame.
+    // 4. Translate the results into a Spark DataFrame.
     val outcomeTranslator = new DeequOutcomeTranslator(df.sparkSession)
-    outcomeTranslator.translate(verificationResult)
+    outcomeTranslator.translate(result)
   }
 
 }
