+++ src/main/scala/com/amazon/deequ/dqdl/execution/DQDLExecutor.scala
@@ -14,22 +14,20 @@
  *
  */
 
-package com.amazon.deequ.dqdl.util
+package com.amazon.deequ.dqdl.execution
 
-import com.amazon.deequ.checks.{Check, CheckWithLastConstraintFilterable}
-import software.amazon.glue.dqdl.model.DQRule
+import com.amazon.deequ.checks.Check
+import com.amazon.deequ.{VerificationResult, VerificationSuite}
+import org.apache.spark.sql.DataFrame
 
-object DataQualityUtility {
-
-  def convertWhereClauseForMetric(whereClause: String): Option[String] =
-    Option(whereClause).map(_ => s"(where: $whereClause)")
-
-  def isWhereClausePresent(rule: DQRule): Boolean = {
-    rule.getWhereClause != null
+/**
+ * Executes a sequence of Deequ Checks on a Spark DataFrame.
+ */
+object DQDLExecutor {
+  def executeRules(df: DataFrame, checks: Seq[Check]): VerificationResult = {
+    VerificationSuite()
+      .onData(df)
+      .addChecks(checks)
+      .run()
   }
-
-  def addWhereClause(rule: DQRule, check: CheckWithLastConstraintFilterable): Check =
-    if (isWhereClausePresent(rule)) check.where(rule.getWhereClause)
-    else check
-
 }
